Model for Hw 4
So I figured out why it was mining per complete  block (1,7), it was because that was the only positive block on the first row so gams was choosing to mine a little of the other blocks (negative blocks) and everything of that positive block. I changed to a negative value so you can see the difference and I also increase the demand to 39 so that gams mines the blocks to full capacity 1.